Draft Bitcoin Improvement Proposals for voting pool HD wallets
These BIPS were submitted to the bitcoin-development mailing list
in order to obtain numbering for the purposes of a BIP-43 purpose
code.

Since there was no response on the mailing list, and therefore no
number assigned, the purpose codes have been changed to ASCII
constants.

These BIPS are still published for informational purposes, even if
they never are accepted into the BIP system.