Draft Bitcoin Improvement Proposal for smart property coloring
This BIP defines a new color kernel for transaction-based coloring.

SPOBC colored coins are undivisible and require no padding.

Since there can be only one unit, the satoshi value is not
relevant.

Smart property coins have a current position and a optional value.

The current position is unspent outpoint containing the color.

The value is up to 40 bytes of OP_RETURN data contained in the
transaction associated with the current position.